requirement.txt
pandas
pyarrow
requests
psycopg2-binary
matplotlib
seaborn